// src/types.ts
export type resourceObj = {
  collectionId: string
  metadata: {
    author: string
    created: number
    modified?: number
  }
  resources: {
    resourceId: string
    resourceCollectionId: string
    resourceStatusCode: number
    resourceAccessorPath: string
    resourceAccessorMethod: 'GET' | 'POST' | 'PUT' | 'DELETE'
    resourceAccessorHeaders: [string, string][]
    resourceStateful: boolean
    resourceHeaders: [string, string][]
    resourceBytes: string
    resourceDescription: string
  }[]
}